Split up and remove storage::mvcc::reader::util
