+++ src/storage/mvcc/reader/mod.rs
@@ -3,7 +3,6 @@
 mod point_getter;
 mod reader;
 mod scanner;
-mod util;
 
 pub use self::point_getter::{PointGetter, PointGetterBuilder};
 pub use self::reader::MvccReader;
