+++ src/storage/mvcc/reader/point_getter.rs
@@ -163,7 +163,7 @@ impl<S: Snapshot> PointGetter<S> {
         if let Some(ref lock_value) = lock_value {
             self.statistics.lock.processed += 1;
             let lock = Lock::parse(lock_value)?;
-            super::util::check_lock(user_key, self.ts, lock)
+            lock.check_ts_conflict(user_key, self.ts)
         } else {
             Ok(())
         }
