+++ src/storage/mvcc/reader/reader.rs
@@ -158,7 +158,7 @@ impl<S: Snapshot> MvccReader<S> {
     /// Returns the blocking lock as the `Err` variant.
     fn check_lock(&mut self, key: &Key, ts: u64) -> Result<()> {
         if let Some(lock) = self.load_lock(key)? {
-            return super::util::check_lock(key, ts, lock);
+            return lock.check_ts_conflict(key, ts);
         }
         Ok(())
     }
