+++ src/storage/mvcc/reader/scanner/backward.rs
@@ -138,8 +138,7 @@ impl<S: Snapshot> BackwardScanner<S> {
                             let lock_value = self.lock_cursor.value(&mut self.statistics.lock);
                             Lock::parse(lock_value)?
                         };
-                        result = super::super::util::check_lock(&current_user_key, ts, lock)
-                            .map(|_| None);
+                        result = lock.check_ts_conflict(&current_user_key, ts).map(|_| None);
                     }
                     IsolationLevel::Rc => {}
                 }
@@ -307,7 +306,7 @@ impl<S: Snapshot> BackwardScanner<S> {
             None => {
                 // Value is in the default CF.
                 self.ensure_default_cursor()?;
-                let value = super::super::util::near_reverse_load_data_by_write(
+                let value = super::near_reverse_load_data_by_write(
                     &mut self.default_cursor.as_mut().unwrap(),
                     user_key,
                     write,
