+++ src/storage/mvcc/reader/scanner/forward.rs
@@ -163,8 +163,7 @@ impl<S: Snapshot> ForwardScanner<S> {
                             let lock_value = self.lock_cursor.value(&mut self.statistics.lock);
                             Lock::parse(lock_value)?
                         };
-                        result = super::super::util::check_lock(&current_user_key, ts, lock)
-                            .map(|_| None);
+                        result = lock.check_ts_conflict(&current_user_key, ts).map(|_| None);
                     }
                     IsolationLevel::Rc => {}
                 }
@@ -298,7 +297,7 @@ impl<S: Snapshot> ForwardScanner<S> {
             None => {
                 // Value is in the default CF.
                 self.ensure_default_cursor()?;
-                let value = super::super::util::near_load_data_by_write(
+                let value = super::near_load_data_by_write(
                     &mut self.default_cursor.as_mut().unwrap(),
                     user_key,
                     write,
