+++ src/storage/types.rs
@@ -97,36 +97,3 @@ pub enum StorageCb {
     Locks(Callback<Vec<LockInfo>>),
     TxnStatus(Callback<TxnStatus>),
 }
-
-#[derive(Clone, Default)]
-pub struct Options {
-    pub lock_ttl: u64,
-    pub skip_constraint_check: bool,
-    pub key_only: bool,
-    pub reverse_scan: bool,
-    pub is_first_lock: bool,
-    pub for_update_ts: u64,
-    pub is_pessimistic_lock: Vec<bool>,
-    // How many keys this transaction involved.
-    pub txn_size: u64,
-    pub min_commit_ts: u64,
-    // Time to wait for lock released in milliseconds when encountering locks.
-    // 0 means using default timeout. Negative means no wait.
-    pub wait_timeout: i64,
-}
-
-impl Options {
-    pub fn new(lock_ttl: u64, skip_constraint_check: bool, key_only: bool) -> Options {
-        Options {
-            lock_ttl,
-            skip_constraint_check,
-            key_only,
-            ..Default::default()
-        }
-    }
-
-    pub fn reverse_scan(mut self) -> Options {
-        self.reverse_scan = true;
-        self
-    }
-}
